Second push of Test changes + namespacing
