+++ code/solr/reindex/handlers/SolrReindexImmediateHandler.php
@@ -1,6 +1,9 @@
 <?php
+
 namespace SilverStripe\FullTextSearch\Solr\Reindex\Handlers;
+
 use Psr\Log\LoggerInterface;
+use SilverStripe\FullTextSearch\Solr\SolrIndex;
 
 /**
  * Invokes an immediate reindex
