+++ code/solr/reindex/handlers/SolrReindexQueuedHandler.php
@@ -1,6 +1,9 @@
 <?php
+
 namespace SilverStripe\FullTextSearch\Solr\Reindex\Handlers;
+
 use Psr\Log\LoggerInterface;
+use SilverStripe\FullTextSearch\Solr\SolrIndex;
 
 if (!interface_exists('QueuedJob')) {
     return;
