+++ tests/SolrIndexSubsitesTest/SolrIndexSubsitesTest_Index.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace SilverStripe\FullTextSearch\Tests\SolrIndexSubsitesTest;
+
+use SilverStripe\FullTextSearch\Solr\SolrIndex;
+
+class SolrIndexSubsitesTest_Index extends SolrIndex
+{
+    public function init()
+    {
+        $this->addClass('File');
+        $this->addClass('SiteTree');
+        $this->addAllFulltextFields();
+    }
+}
