+++ tests/SolrReindexTest.php
@@ -1,13 +1,6 @@
 <?php
 
-use Monolog\Handler\AbstractProcessingHandler;
-use Monolog\Handler\HandlerInterface;
-use Monolog\Logger;
-use Psr\Log\LoggerInterface;
 use SilverStripe\Dev\SapphireTest;
-use SilverStripe\Dev\TestOnly;
-use SilverStripe\ORM\DataObject;
-use SilverStripe\ORM\DataExtension;
 
 if (class_exists('Phockito')) {
     Phockito::include_hamcrest(false);
@@ -285,282 +278,5 @@ class SolrReindexTest extends SapphireTest
         Phockito::verify($this->service, 1)->deleteByQuery(
             '+(ClassHierarchy:SolrReindexTest_Item) +_query_:"{!frange l=5 u=5}mod(ID, 6)" +(_testvariant:"1")'
         );
-    }
-}
-
-/**
- * Provides a wrapper for testing SolrReindexBase
- */
-class SolrReindexTest_TestHandler extends SolrReindexBase
-{
-    public function processGroup(
-        LoggerInterface $logger, SolrIndex $indexInstance, $state, $class, $groups, $group, $taskName
-    ) {
-        $indexName = $indexInstance->getIndexName();
-        $stateName = json_encode($state);
-        $logger->info("Called processGroup with {$indexName}, {$stateName}, {$class}, group {$group} of {$groups}");
-    }
-
-    public function triggerReindex(LoggerInterface $logger, $batchSize, $taskName, $classes = null)
-    {
-        $logger->info("Called triggerReindex");
-    }
-}
-
-
-class SolrReindexTest_Index extends SolrIndex implements TestOnly
-{
-    public function init()
-    {
-        $this->addClass('SolrReindexTest_Item');
-        $this->addAllFulltextFields();
-    }
-}
-
-/**
- * Does not have any variant extensions
- */
-class SolrReindexTest_Item extends DataObject implements TestOnly
-{
-    private static $extensions = array(
-        'SolrReindexTest_ItemExtension'
-    );
-
-    private static $db = array(
-        'Title' => 'Varchar(255)',
-        'Variant' => 'Int(0)'
-    );
-}
-
-/**
- * Select only records in the current variant
- */
-class SolrReindexTest_ItemExtension extends DataExtension implements TestOnly
-{
-    /**
-     * Filter records on the current variant
-     *
-     * @param SQLQuery $query
-     * @param DataQuery $dataQuery
-     */
-    public function augmentSQL(SilverStripe\ORM\Queries\SQLSelect $query, SilverStripe\ORM\DataQuery $dataQuery = NULL)
-    {
-        $variant = SolrReindexTest_Variant::get_current();
-        if ($variant !== null && !$query->filtersOnID()) {
-            $sqlVariant = Convert::raw2sql($variant);
-            $query->addWhere("\"Variant\" = '{$sqlVariant}'");
-        }
-    }
-}
-
-
-/**
- * Dummy variant that selects items with field Varient matching the current value
- *
- * Variant states are 0 and 1, or null if disabled
- */
-class SolrReindexTest_Variant extends SearchVariant implements TestOnly
-{
-    /**
-     * Value of this variant (either null, 0, or 1)
-     *
-     * @var int|null
-     */
-    protected static $current = null;
-
-    /**
-     * Activate this variant
-     */
-    public static function enable()
-    {
-        self::disable();
-
-        self::$current = 0;
-        self::$variants = array(
-            'SolrReindexTest_Variant' => singleton('SolrReindexTest_Variant')
-        );
-    }
-
-    /**
-     * Disable this variant and reset
-     */
-    public static function disable()
-    {
-        self::$current = null;
-        self::$variants = null;
-        self::$class_variants = array();
-        self::$call_instances = array();
-    }
-
-    public function activateState($state)
-    {
-        self::set_current($state);
-    }
-
-    /**
-     * Set the current variant to the given state
-     *
-     * @param int $current 0, 1, 2, or null (disabled)
-     */
-    public static function set_current($current)
-    {
-        self::$current = $current;
-    }
-
-    /**
-     * Get the current state
-     *
-     * @return string|null
-     */
-    public static function get_current()
-    {
-        // Always use string values for states for consistent json_encode value
-        if (isset(self::$current)) {
-            return (string)self::$current;
-        }
-    }
-
-    public function alterDefinition($class, $index)
-    {
-        $self = get_class($this);
-
-        $this->addFilterField($index, '_testvariant', array(
-            'name' => '_testvariant',
-            'field' => '_testvariant',
-            'fullfield' => '_testvariant',
-            'base' => ClassInfo::baseDataClass($class),
-            'origin' => $class,
-            'type' => 'Int',
-            'lookup_chain' => array(array('call' => 'variant', 'variant' => $self, 'method' => 'currentState'))
-        ));
-    }
-
-    public function alterQuery($query, $index)
-    {
-        // I guess just calling it _testvariant is ok?
-        $query->filter('_testvariant', $this->currentState());
-    }
-
-    public function appliesTo($class, $includeSubclasses)
-    {
-        return $class === 'SolrReindexTest_Item' ||
-            ($includeSubclasses && is_subclass_of($class, 'SolrReindexTest_Item', true));
-    }
-
-    public function appliesToEnvironment()
-    {
-        // Set to null to disable
-        return self::$current !== null;
-    }
-
-    public function currentState()
-    {
-        return self::get_current();
-    }
-
-    public function reindexStates()
-    {
-        // Always use string values for states for consistent json_encode value
-        return array('0', '1', '2');
-    }
-}
-
-/**
- * Test logger for recording messages
- */
-class SolrReindexTest_RecordingLogger extends Logger implements TestOnly
-{
-    /**
-     * @var SolrReindexTest_Handler
-     */
-    protected $testHandler = null;
-
-    public function __construct($name = 'testlogger', array $handlers = array(), array $processors = array())
-    {
-        parent::__construct($name, $handlers, $processors);
-
-        $this->testHandler = new SolrReindexTest_Handler();
-        $this->pushHandler($this->testHandler);
-    }
-
-    /**
-     * @return array
-     */
-    public function getMessages()
-    {
-        return $this->testHandler->getMessages();
-    }
-
-    /**
-     * Clear all messages
-     */
-    public function clear()
-    {
-        $this->testHandler->clear();
-    }
-
-    /**
-     * Get messages with the given filter
-     *
-     * @param string $containing
-     * @return array Filtered array
-     */
-    public function filterMessages($containing)
-    {
-        return array_values(array_filter(
-            $this->getMessages(),
-            function ($content) use ($containing) {
-                return stripos($content, $containing) !== false;
-            }
-        ));
-    }
-
-    /**
-     * Count all messages containing the given substring
-     *
-     * @param string $containing Message to filter by
-     * @return int
-     */
-    public function countMessages($containing = null)
-    {
-        if ($containing) {
-            $messages = $this->filterMessages($containing);
-        } else {
-            $messages = $this->getMessages();
-        }
-        return count($messages);
-    }
-}
-
-/**
- * Logger for recording messages for later retrieval
- */
-class SolrReindexTest_Handler extends AbstractProcessingHandler implements TestOnly
-{
-    /**
-     * Messages
-     *
-     * @var array
-     */
-    protected $messages = array();
-
-    /**
-     * Get all messages
-     *
-     * @return array
-     */
-    public function getMessages()
-    {
-        return $this->messages;
-    }
-
-    public function clear()
-    {
-        $this->messages = array();
-    }
-
-    protected function write(array $record)
-    {
-        $this->messages[] = $record['message'];
     }
 }
